Refactor more of the filter definitions into their matching header
